example_project/requirements.txt
Django==1.4.2
git+https://github.com/nosamanuel/django-hstore.git@hstore-ops
psycopg2==2.4.5
django-jsonfield==0.8.11
csvkit==0.5.0
django-tastypie==0.9.11
